Minor cleanup Add docstring
